requirements_cpu.txt
easydict==1.9
future==0.17.1
gast==0.2.2
gin-config==0.1.3
gym==0.14.0
h5py==2.9.0
matplotlib==2.2.4
notebook==5.7.8
numpy==1.16.5
PyYAML==5.1.2
scipy==1.2.2
six==1.11.0
tf-nightly==1.15.0.dev20190821
tf-agents-nightly==0.2.0.dev20190831
tb-nightly==1.15.0a20190831
tf-estimator-nightly==1.14.0.dev2019083101
tfp-nightly==0.9.0.dev20190831
